+++ docs/tr1/CHANGELOG.md
@@ -12,6 +12,7 @@
 - added optional dynamic lighting for gun flashes and explosions, similar to TR2+ (#2357)
 - added a `/cut` (alias: `/cutscene`) console command for playing cutscenes
 - added a `/gym` (alias: `/home`) console command for playing Lara's Home
+- added a `/music` console command that plays a specific music track
 - added a console log when using the `/demo` command
 - ⚠️ changed the game data to use a separate strings file for text information, removing it from the game flow file
 - changed the object texture limit from 2048 to unlimited (within game's overall memory cap)