+++ docs/tr2/CHANGELOG.md
@@ -10,6 +10,7 @@
 - added support for animated sprites (#2401)
 - added a `/cut` (alias: `/cutscene`) console command for playing cutscenes
 - added a `/gym` (alias: `/home`) console command for playing Lara's Home
+- added a `/music` console command that plays a specific music track
 - added a console log when using the `/demo` command
 - changed the object texture limit from 2048 to unlimited (within game's overall memory cap) (#1795)
 - changed the sprite texture limit from 512 to unlimited (within game's overall memory cap) (#1795)